feat(service,pdf): add printnode options to generation logs and change print_node_job_id to bigint

+++ internal/repository/pdf.go
@@ -137,7 +137,7 @@ func (r *PDFRepository) InsertPrintLog(
 	ctx context.Context,
 	exec db.PGExecutor,
 	input model.CreatePDFPrintLogParams,
-	jobID *int,
+	jobID *int64,
 	errorMessage *string,
 ) (int, error) {
 	var id int
